back-end/requirements.txt
flask
requests
flask_cors
python-dotenv
psycopg2-binary
countryinfo